+++ app/api/events.py
@@ -8,7 +8,6 @@ from flask_rest_jsonapi.exceptions import ObjectNotFound
 from marshmallow_jsonapi import fields
 from marshmallow_jsonapi.flask import Schema
 from sqlalchemy import and_, or_
-from sqlalchemy.orm.exc import NoResultFound
 
 from app.api.bootstrap import api
 from app.api.data_layers.EventCopyLayer import EventCopyLayer
@@ -61,6 +60,7 @@ from app.models.user import (
 )
 from app.models.user_favourite_event import UserFavouriteEvent
 from app.models.users_events_role import UsersEventsRoles
+from app.models.video_stream import VideoStream
 
 
 def validate_event(user, data):
@@ -355,267 +355,59 @@ class EventList(ResourceList):
     }
 
 
+def set_event_id(model, identifier, kwargs, attr='event_id', column_name='id'):
+    if kwargs.get('id'):  # ID already set
+        return
+    if kwargs.get(identifier) is None:
+        return
+    item = safe_query_kwargs(model, kwargs, identifier, column_name=column_name)
+    kwargs['id'] = getattr(item, attr, None)
+
+
 def get_id(view_kwargs):
     """
     method to get the resource id for fetching details
     :param view_kwargs:
     :return:
     """
-    if view_kwargs.get('identifier'):
-        event = safe_query_kwargs(Event, view_kwargs, 'identifier', 'identifier')
-        view_kwargs['id'] = event.id
+    set_event_id(Event, 'identifier', view_kwargs, attr='id', column_name='identifier')
 
-    if view_kwargs.get('sponsor_id') is not None:
-        sponsor = safe_query_kwargs(Sponsor, view_kwargs, 'sponsor_id')
-        if sponsor.event_id is not None:
-            view_kwargs['id'] = sponsor.event_id
-        else:
-            view_kwargs['id'] = None
+    lookup_list = [
+        (Sponsor, 'sponsor_id'),
+        (UserFavouriteEvent, 'user_favourite_event_id'),
+        (EventCopyright, 'copyright_id'),
+        (Track, 'track_id'),
+        (SessionType, 'session_type_id'),
+        (FaqType, 'faq_type_id'),
+        (EventInvoice, 'event_invoice_id'),
+        (DiscountCode, 'discount_code_id'),
+        (Session, 'session_id'),
+        (SocialLink, 'social_link_id'),
+        (Tax, 'tax_id'),
+        (StripeAuthorization, 'stripe_authorization_id'),
+        (DiscountCode, 'discount_code_id'),
+        (SpeakersCall, 'speakers_call_id'),
+        (Ticket, 'ticket_id'),
+        (TicketTag, 'ticket_tag_id'),
+        (RoleInvite, 'role_invite_id'),
+        (UsersEventsRoles, 'users_events_role_id'),
+        (AccessCode, 'access_code_id'),
+        (Speaker, 'speaker_id'),
+        (EmailNotification, 'email_notification_id'),
+        (Microlocation, 'microlocation_id'),
+        (TicketHolder, 'attendee_id'),
+        (CustomForms, 'custom_form_id'),
+        (Faq, 'faq_id'),
+        (Feedback, 'feedback_id'),
+        (VideoStream, 'video_stream_id'),
+    ]
+    for model, identifier in lookup_list:
+        set_event_id(model, identifier, view_kwargs)
 
-    if view_kwargs.get('user_favourite_event_id') is not None:
-        user_favourite_event = safe_query_kwargs(
-            UserFavouriteEvent,
-            view_kwargs,
-            'user_favourite_event_id',
-        )
-        if user_favourite_event.event_id is not None:
-            view_kwargs['id'] = user_favourite_event.event_id
-        else:
-            view_kwargs['id'] = None
-
-    if view_kwargs.get('copyright_id') is not None:
-        copyright = safe_query_kwargs(EventCopyright, view_kwargs, 'copyright_id')
-        if copyright.event_id is not None:
-            view_kwargs['id'] = copyright.event_id
-        else:
-            view_kwargs['id'] = None
-
-    if view_kwargs.get('track_id') is not None:
-        track = safe_query_kwargs(Track, view_kwargs, 'track_id')
-        if track.event_id is not None:
-            view_kwargs['id'] = track.event_id
-        else:
-            view_kwargs['id'] = None
-
-    if view_kwargs.get('session_type_id') is not None:
-        session_type = safe_query_kwargs(SessionType, view_kwargs, 'session_type_id')
-        if session_type.event_id is not None:
-            view_kwargs['id'] = session_type.event_id
-        else:
-            view_kwargs['id'] = None
-
-    if view_kwargs.get('faq_type_id') is not None:
-        faq_type = safe_query_kwargs(FaqType, view_kwargs, 'faq_type_id')
-        if faq_type.event_id is not None:
-            view_kwargs['id'] = faq_type.event_id
-        else:
-            view_kwargs['id'] = None
-
-    if view_kwargs.get('event_invoice_id') is not None:
-        event_invoice = safe_query_kwargs(EventInvoice, view_kwargs, 'event_invoice_id')
-        if event_invoice.event_id is not None:
-            view_kwargs['id'] = event_invoice.event_id
-        else:
-            view_kwargs['id'] = None
-
-    if view_kwargs.get('event_invoice_identifier') is not None:
-        event_invoice = safe_query_kwargs(
-            EventInvoice, view_kwargs, 'event_invoice_identifier', 'identifier'
-        )
-        if event_invoice.event_id is not None:
-            view_kwargs['id'] = event_invoice.event_id
-        else:
-            view_kwargs['id'] = None
-
-    if view_kwargs.get('discount_code_id') is not None:
-        discount_code = safe_query_kwargs(DiscountCode, view_kwargs, 'discount_code_id')
-        if discount_code.event_id is not None:
-            view_kwargs['id'] = discount_code.event_id
-        else:
-            view_kwargs['id'] = None
-
-    if view_kwargs.get('session_id') is not None:
-        sessions = safe_query_kwargs(Session, view_kwargs, 'session_id')
-        if sessions.event_id is not None:
-            view_kwargs['id'] = sessions.event_id
-        else:
-            view_kwargs['id'] = None
-
-    if view_kwargs.get('social_link_id') is not None:
-        social_link = safe_query_kwargs(SocialLink, view_kwargs, 'social_link_id')
-        if social_link.event_id is not None:
-            view_kwargs['id'] = social_link.event_id
-        else:
-            view_kwargs['id'] = None
-
-    if view_kwargs.get('tax_id') is not None:
-        tax = safe_query_kwargs(Tax, view_kwargs, 'tax_id')
-        if tax.event_id is not None:
-            view_kwargs['id'] = tax.event_id
-        else:
-            view_kwargs['id'] = None
-
-    if view_kwargs.get('stripe_authorization_id') is not None:
-        stripe_authorization = safe_query_kwargs(
-            StripeAuthorization,
-            view_kwargs,
-            'stripe_authorization_id',
-        )
-        if stripe_authorization.event_id is not None:
-            view_kwargs['id'] = stripe_authorization.event_id
-        else:
-            view_kwargs['id'] = None
-
-    if view_kwargs.get('user_id') is not None:
-        try:
-            discount_code = (
-                db.session.query(DiscountCode)
-                .filter_by(id=view_kwargs['discount_code_id'])
-                .one()
-            )
-        except NoResultFound:
-            raise ObjectNotFound(
-                {'parameter': 'discount_code_id'},
-                "DiscountCode: {} not found".format(view_kwargs['discount_code_id']),
-            )
-        else:
-            if discount_code.event_id is not None:
-                view_kwargs['id'] = discount_code.event_id
-            else:
-                view_kwargs['id'] = None
-
-    if view_kwargs.get('speakers_call_id') is not None:
-        speakers_call = safe_query_kwargs(SpeakersCall, view_kwargs, 'speakers_call_id')
-        if speakers_call.event_id is not None:
-            view_kwargs['id'] = speakers_call.event_id
-        else:
-            view_kwargs['id'] = None
-
-    if view_kwargs.get('ticket_id') is not None:
-        ticket = safe_query_kwargs(Ticket, view_kwargs, 'ticket_id')
-        if ticket.event_id is not None:
-            view_kwargs['id'] = ticket.event_id
-        else:
-            view_kwargs['id'] = None
-
-    if view_kwargs.get('ticket_tag_id') is not None:
-        ticket_tag = safe_query_kwargs(TicketTag, view_kwargs, 'ticket_tag_id')
-        if ticket_tag.event_id is not None:
-            view_kwargs['id'] = ticket_tag.event_id
-        else:
-            view_kwargs['id'] = None
-
-    if view_kwargs.get('role_invite_id') is not None:
-        role_invite = safe_query_kwargs(RoleInvite, view_kwargs, 'role_invite_id')
-        if role_invite.event_id is not None:
-            view_kwargs['id'] = role_invite.event_id
-        else:
-            view_kwargs['id'] = None
-
-    if view_kwargs.get('users_events_role_id') is not None:
-        users_events_role = safe_query_kwargs(
-            UsersEventsRoles,
-            view_kwargs,
-            'users_events_role_id',
-        )
-        if users_events_role.event_id is not None:
-            view_kwargs['id'] = users_events_role.event_id
-
-    if view_kwargs.get('access_code_id') is not None:
-        access_code = safe_query_kwargs(AccessCode, view_kwargs, 'access_code_id')
-        if access_code.event_id is not None:
-            view_kwargs['id'] = access_code.event_id
-        else:
-            view_kwargs['id'] = None
-
-    if view_kwargs.get('speaker_id'):
-        try:
-            speaker = (
-                db.session.query(Speaker).filter_by(id=view_kwargs['speaker_id']).one()
-            )
-        except NoResultFound:
-            raise ObjectNotFound(
-                {'parameter': 'speaker_id'},
-                "Speaker: {} not found".format(view_kwargs['speaker_id']),
-            )
-        else:
-            if speaker.event_id:
-                view_kwargs['id'] = speaker.event_id
-            else:
-                view_kwargs['id'] = None
-
-    if view_kwargs.get('email_notification_id'):
-        try:
-            email_notification = (
-                db.session.query(EmailNotification)
-                .filter_by(id=view_kwargs['email_notification_id'])
-                .one()
-            )
-        except NoResultFound:
-            raise ObjectNotFound(
-                {'parameter': 'email_notification_id'},
-                "Email Notification: {} not found".format(
-                    view_kwargs['email_notification_id']
-                ),
-            )
-        else:
-            if email_notification.event_id:
-                view_kwargs['id'] = email_notification.event_id
-            else:
-                view_kwargs['id'] = None
-
-    if view_kwargs.get('microlocation_id'):
-        try:
-            microlocation = (
-                db.session.query(Microlocation)
-                .filter_by(id=view_kwargs['microlocation_id'])
-                .one()
-            )
-        except NoResultFound:
-            raise ObjectNotFound(
-                {'parameter': 'microlocation_id'},
-                "Microlocation: {} not found".format(view_kwargs['microlocation_id']),
-            )
-        else:
-            if microlocation.event_id:
-                view_kwargs['id'] = microlocation.event_id
-            else:
-                view_kwargs['id'] = None
-
-    if view_kwargs.get('attendee_id'):
-        attendee = safe_query_kwargs(TicketHolder, view_kwargs, 'attendee_id')
-        if attendee.event_id is not None:
-            view_kwargs['id'] = attendee.event_id
-        else:
-            view_kwargs['id'] = None
-
-    if view_kwargs.get('custom_form_id') is not None:
-        custom_form = safe_query_kwargs(CustomForms, view_kwargs, 'custom_form_id')
-        if custom_form.event_id is not None:
-            view_kwargs['id'] = custom_form.event_id
-        else:
-            view_kwargs['id'] = None
-
-    if view_kwargs.get('faq_id') is not None:
-        faq = safe_query_kwargs(Faq, view_kwargs, 'faq_id')
-        if faq.event_id is not None:
-            view_kwargs['id'] = faq.event_id
-        else:
-            view_kwargs['id'] = None
-
-    if view_kwargs.get('order_identifier') is not None:
-        order = safe_query_kwargs(Order, view_kwargs, 'order_identifier', 'identifier')
-        if order.event_id is not None:
-            view_kwargs['id'] = order.event_id
-        else:
-            view_kwargs['id'] = None
-
-    if view_kwargs.get('feedback_id') is not None:
-        feedback = safe_query_kwargs(Feedback, view_kwargs, 'feedback_id')
-        if feedback.event_id is not None:
-            view_kwargs['id'] = feedback.event_id
-        else:
-            view_kwargs['id'] = None
+    set_event_id(
+        EventInvoice, 'event_invoice_identifier', view_kwargs, column_name='identifier'
+    )
+    set_event_id(Order, 'order_identifier', view_kwargs, column_name='identifier')
 
     return view_kwargs
 
@@ -646,17 +438,8 @@ class EventDetail(ResourceDetail):
         """
         get_id(view_kwargs)
 
-        if view_kwargs.get('order_identifier') is not None:
-            order = safe_query_kwargs(
-                Order, view_kwargs, 'order_identifier', 'identifier'
-            )
-            if order.event_id is not None:
-                view_kwargs['id'] = order.event_id
-            else:
-                view_kwargs['id'] = None
-
     def after_get_object(self, event, view_kwargs):
-        if event.state == "draft":
+        if event and event.state == "draft":
             if not is_logged_in() or not has_access('is_coorganizer', event_id=event.id):
                 raise ObjectNotFound({'parameter': '{id}'}, "Event: not found")