+++ app/api/schema/events.py
@@ -8,7 +8,7 @@ from sqlalchemy.orm.exc import NoResultFound
 
 from app.api.helpers.errors import UnprocessableEntityError
 from app.api.helpers.utilities import dasherize
-from app.api.schema.base import SoftDeletionSchema
+from app.api.schema.base import GetterRelationship, SoftDeletionSchema
 from app.models.event import Event
 
 
@@ -102,7 +102,6 @@ class EventSchemaPublic(SoftDeletionSchema):
     is_stripe_linked = fields.Boolean(dump_only=True, allow_none=True, default=False)
 
     tickets = Relationship(
-        attribute='tickets',
         self_view='v1.event_ticket',
         self_view_kwargs={'id': '<id>'},
         related_view='v1.ticket_list',
@@ -112,7 +111,6 @@ class EventSchemaPublic(SoftDeletionSchema):
         type_='ticket',
     )
     faqs = Relationship(
-        attribute='faqs',
         self_view='v1.event_faqs',
         self_view_kwargs={'id': '<id>'},
         related_view='v1.faq_list',
@@ -122,7 +120,6 @@ class EventSchemaPublic(SoftDeletionSchema):
         type_='faq',
     )
     faq_types = Relationship(
-        attribute='faq_types',
         self_view='v1.event_faq_types',
         self_view_kwargs={'id': '<id>'},
         related_view='v1.faq_type_list',
@@ -132,7 +129,6 @@ class EventSchemaPublic(SoftDeletionSchema):
         type_='faq_type',
     )
     feedbacks = Relationship(
-        attribute='feedbacks',
         self_view='v1.event_feedbacks',
         self_view_kwargs={'id': '<id>'},
         related_view='v1.feedback_list',
@@ -142,7 +138,6 @@ class EventSchemaPublic(SoftDeletionSchema):
         type_='feedback',
     )
     ticket_tags = Relationship(
-        attribute='ticket_tags',
         self_view='v1.event_ticket_tag',
         self_view_kwargs={'id': '<id>'},
         related_view='v1.ticket_tag_list',
@@ -192,7 +187,6 @@ class EventSchemaPublic(SoftDeletionSchema):
         type_='sponsor',
     )
     speakers_call = Relationship(
-        attribute='speakers_call',
         self_view='v1.event_speakers_call',
         self_view_kwargs={'id': '<id>'},
         related_view='v1.speakers_call_detail',
@@ -220,7 +214,6 @@ class EventSchemaPublic(SoftDeletionSchema):
         type_='event-copyright',
     )
     tax = Relationship(
-        attribute='tax',
         self_view='v1.event_tax',
         self_view_kwargs={'id': '<id>'},
         related_view='v1.tax_detail',
@@ -249,7 +242,6 @@ class EventSchemaPublic(SoftDeletionSchema):
         type_='speaker',
     )
     event_type = Relationship(
-        attribute='event_type',
         self_view='v1.event_event_type',
         self_view_kwargs={'id': '<id>'},
         related_view='v1.event_type_detail',
@@ -258,7 +250,6 @@ class EventSchemaPublic(SoftDeletionSchema):
         type_='event-type',
     )
     event_topic = Relationship(
-        attribute='event_topic',
         self_view='v1.event_event_topic',
         self_view_kwargs={'id': '<id>'},
         related_view='v1.event_topic_detail',
@@ -267,7 +258,6 @@ class EventSchemaPublic(SoftDeletionSchema):
         type_='event-topic',
     )
     event_sub_topic = Relationship(
-        attribute='event_sub_topic',
         self_view='v1.event_event_sub_topic',
         self_view_kwargs={'id': '<id>'},
         related_view='v1.event_sub_topic_detail',
@@ -286,7 +276,6 @@ class EventSchemaPublic(SoftDeletionSchema):
         type_='custom-form',
     )
     owner = Relationship(
-        attribute='owner',
         self_view='v1.event_owner',
         self_view_kwargs={'id': '<id>'},
         related_view='v1.user_detail',
@@ -295,7 +284,6 @@ class EventSchemaPublic(SoftDeletionSchema):
         type_='user',
     )
     organizers = Relationship(
-        attribute='organizers',
         self_view='v1.event_organizers',
         self_view_kwargs={'id': '<id>'},
         related_view='v1.user_list',
@@ -304,7 +292,6 @@ class EventSchemaPublic(SoftDeletionSchema):
         many=True,
     )
     coorganizers = Relationship(
-        attribute='coorganizers',
         self_view='v1.event_coorganizers',
         self_view_kwargs={'id': '<id>'},
         related_view='v1.user_list',
@@ -313,7 +300,6 @@ class EventSchemaPublic(SoftDeletionSchema):
         many=True,
     )
     stripe_authorization = Relationship(
-        attribute='stripe_authorization',
         self_view='v1.stripe_authorization_event',
         self_view_kwargs={'id': '<id>'},
         related_view='v1.stripe_authorization_detail',
@@ -336,6 +322,15 @@ class EventSchemaPublic(SoftDeletionSchema):
         related_view_kwargs={'id': '<id>'},
         schema='EventStatisticsGeneralSchema',
         type_='event-statistics-general',
+    )
+    video_stream = GetterRelationship(
+        getter='safe_video_stream',
+        self_view='v1.video_stream_event',
+        self_view_kwargs={'id': '<id>'},
+        related_view='v1.video_stream_detail',
+        related_view_kwargs={'event_id': '<id>'},
+        schema='VideoStreamSchema',
+        type_='video-stream',
     )
 
 
@@ -367,7 +362,6 @@ class EventSchema(EventSchemaPublic):
         type_='discount-code',
     )
     track_organizers = Relationship(
-        attribute='track_organizers',
         self_view='v1.event_track_organizers',
         self_view_kwargs={'id': '<id>'},
         related_view='v1.user_list',
@@ -376,7 +370,6 @@ class EventSchema(EventSchemaPublic):
         many=True,
     )
     moderators = Relationship(
-        attribute='moderators',
         self_view='v1.event_moderators',
         self_view_kwargs={'id': '<id>'},
         related_view='v1.user_list',
@@ -385,7 +378,6 @@ class EventSchema(EventSchemaPublic):
         many=True,
     )
     registrars = Relationship(
-        attribute='registrars',
         self_view='v1.event_registrars',
         self_view_kwargs={'id': '<id>'},
         related_view='v1.user_list',
@@ -394,7 +386,6 @@ class EventSchema(EventSchemaPublic):
         many=True,
     )
     orders = Relationship(
-        attribute='orders',
         self_view='v1.event_orders',
         self_view_kwargs={'id': '<id>'},
         related_view='v1.orders_list',
@@ -404,7 +395,6 @@ class EventSchema(EventSchemaPublic):
         many=True,
     )
     role_invites = Relationship(
-        attribute='role_invites',
         self_view='v1.event_role_invite',
         self_view_kwargs={'id': '<id>'},
         related_view='v1.role_invite_list',
@@ -413,7 +403,6 @@ class EventSchema(EventSchemaPublic):
         type_='role-invite',
     )
     access_codes = Relationship(
-        attribute='access_codes',
         self_view='v1.event_access_codes',
         self_view_kwargs={'id': '<id>'},
         related_view='v1.access_code_list',
@@ -423,7 +412,6 @@ class EventSchema(EventSchemaPublic):
         type_='access-code',
     )
     attendees = Relationship(
-        attribute='attendees',
         self_view='v1.event_attendees',
         self_view_kwargs={'id': '<id>'},
         related_view='v1.attendee_list',