+++ app/models/setting.py
@@ -168,6 +168,7 @@ class Setting(db.Model):
     admin_billing_zip = db.Column(db.String)
     admin_billing_additional_info = db.Column(db.String)
     admin_billing_paypal_email = db.Column(db.String)
+    admin_billing_logo = db.Column(db.String)
     #
     # Generators
     #